requirements.txt
streamlit>=1.28.0
pyyaml>=6.0

# Testing dependencies
pytest>=7.4.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0
